pull

+++ exporting/export_package.py
@@ -24,9 +24,11 @@ def export_package(client, args):
 
     access = args.access
     threat = args.threat
+    nat = args.nat
     if args.all:
         access = True
         threat = True
+        nat = True
 
     data_dict = {}
     unexportable_objects = {}
@@ -54,17 +56,18 @@ def export_package(client, args):
                 os.remove(layer_tar_name)
 
         # NAT policy should be exported as a part of Access policy
-        if show_package.data["nat-policy"]:
-            debug_log("Exporting NAT policy", True)
-            nat_data_dict, nat_unexportable_objects = export_nat_rulebase(show_package.data["name"], client)
-            if nat_data_dict:
-                nat_tar_name = "exported__nat_layer__" + show_package.data["name"] + "__" + timestamp + ".tar.gz"
-                with tarfile.open(nat_tar_name, "w:gz") as tar:
-                    export_to_tar(nat_data_dict, timestamp, tar, ["nat-rule", "nat-section"], client.api_version)
-                merge_data(data_dict, nat_data_dict)
-                merge_data(unexportable_objects, nat_unexportable_objects)
-                tar_file.add(nat_tar_name)
-                os.remove(nat_tar_name)
+        if nat:
+            if show_package.data["nat-policy"]:
+                debug_log("Exporting NAT policy", True)
+                nat_data_dict, nat_unexportable_objects = export_nat_rulebase(show_package.data["name"], client)
+                if nat_data_dict:
+                    nat_tar_name = "exported__nat_layer__" + show_package.data["name"] + "__" + timestamp + ".tar.gz"
+                    with tarfile.open(nat_tar_name, "w:gz") as tar:
+                        export_to_tar(nat_data_dict, timestamp, tar, ["nat-rule", "nat-section"], client.api_version)
+                    merge_data(data_dict, nat_data_dict)
+                    merge_data(unexportable_objects, nat_unexportable_objects)
+                    tar_file.add(nat_tar_name)
+                    os.remove(nat_tar_name)
 
     if threat:
         if show_package.data["threat-prevention"]:
